+++ lua/codecompanion/strategies/chat/init.lua
@@ -11,6 +11,8 @@ local util = require("codecompanion.utils")
 local yaml = require("codecompanion.utils.yaml")
 
 local api = vim.api
+local get_node_text = vim.treesitter.get_node_text --[[@type function]]
+local get_query = vim.treesitter.query.get --[[@type function]]
 
 local CONSTANTS = {
   AUTOCMD_GROUP = "codecompanion.chat",
@@ -32,7 +34,8 @@ local function make_id(val)
   return hash.hash(val)
 end
 
-local _cached_settings
+local _cached_settings = {}
+local _yaml_parser
 
 ---Parse the chat buffer for settings
 ---@param bufnr integer
@@ -47,21 +50,29 @@ local function ts_parse_settings(bufnr, adapter)
   if not config.display.chat.show_settings then
     if adapter then
       _cached_settings[bufnr] = adapter:get_default_settings()
-
       return _cached_settings[bufnr]
     end
   end
 
   local settings = {}
-  local parser = vim.treesitter.get_parser(bufnr, "yaml", { ignore_injections = false })
-  local query = vim.treesitter.query.get("yaml", "chat")
-  local root = parser:parse()[1]:root()
+  if not _yaml_parser then
+    _yaml_parser = vim.treesitter.get_parser(bufnr, "yaml", { ignore_injections = false })
+  end
 
-  for _, matches, _ in query:iter_matches(root, bufnr) do
+  local query = get_query("yaml", "chat")
+  local root = _yaml_parser:parse()[1]:root()
+
+  local end_line = -1
+  if adapter then
+    -- Account for the two YAML lines and the fact Tree-sitter is 0-indexed
+    end_line = vim.tbl_count(adapter:get_default_settings()) + 2 - 1
+  end
+
+  for _, matches, _ in query:iter_matches(root, bufnr, 0, end_line) do
     local nodes = matches[1]
     local node = type(nodes) == "table" and nodes[1] or nodes
 
-    local value = vim.treesitter.get_node_text(node, bufnr)
+    local value = get_node_text(node, bufnr)
 
     settings = yaml.decode(value)
     break
@@ -77,11 +88,10 @@ end
 
 ---Parse the chat buffer for the last message
 ---@param chat CodeCompanion.Chat
----@param role string
 ---@param start_range number
----@return table{content: string}
-local function ts_parse_messages(chat, role, start_range)
-  local query = vim.treesitter.query.get("markdown", "chat")
+---@return { content: string }
+local function ts_parse_messages(chat, start_range)
+  local query = get_query("markdown", "chat")
 
   local tree = chat.parser:parse({ start_range - 1, -1 })[1]
   local root = tree:root()
@@ -91,9 +101,9 @@ local function ts_parse_messages(chat, role, start_range)
 
   for id, node in query:iter_captures(root, chat.bufnr, start_range - 1, -1) do
     if query.captures[id] == "role" then
-      last_role = vim.treesitter.get_node_text(node, chat.bufnr)
+      last_role = get_node_text(node, chat.bufnr)
     elseif last_role == chat.ui:format_header(user_role) and query.captures[id] == "content" then
-      table.insert(content, vim.treesitter.get_node_text(node, chat.bufnr))
+      table.insert(content, get_node_text(node, chat.bufnr))
     end
   end
 
@@ -111,7 +121,7 @@ end
 ---@return TSNode | nil
 local function ts_parse_codeblock(chat, cursor)
   local root = chat.parser:parse()[1]:root()
-  local query = vim.treesitter.query.get("markdown", "chat")
+  local query = get_query("markdown", "chat")
   if query == nil then
     return nil
   end
@@ -185,7 +195,7 @@ function Chat.new(args)
   end
   self.parser = parser
 
-  self.References = require("codecompanion.strategies.chat.references").new({ chat = self })
+  self.references = require("codecompanion.strategies.chat.references").new({ chat = self })
   self.watchers = require("codecompanion.strategies.chat.watchers").new()
   self.tools = require("codecompanion.strategies.chat.tools").new({ bufnr = self.bufnr, messages = self.messages })
   self.variables = require("codecompanion.strategies.chat.variables").new()
@@ -315,7 +325,7 @@ function Chat:set_autocmds()
         if errors and node then
           for child in node:iter_children() do
             assert(child:type() == "block_mapping_pair")
-            local key = vim.treesitter.get_node_text(child:named_child(0), self.bufnr)
+            local key = get_node_text(child:named_child(0), self.bufnr)
             if errors[key] then
               local lnum, col, end_lnum, end_col = child:range()
               table.insert(items, {
@@ -350,7 +360,7 @@ function Chat:_get_settings_key(opts)
     return
   end
   local key_node = node:named_child(0)
-  local key_name = vim.treesitter.get_node_text(key_node, self.bufnr)
+  local key_name = get_node_text(key_node, self.bufnr)
   return key_name, node
 end
 
@@ -410,13 +420,14 @@ function Chat:complete_models(request, callback)
 end
 
 ---Set the system prompt in the chat buffer
+---@prompt? string
 ---@return CodeCompanion.Chat
-function Chat:set_system_prompt()
+function Chat:set_system_prompt(prompt)
   if self.opts and self.opts.ignore_system_prompt then
     return self
   end
 
-  local prompt = config.opts.system_prompt
+  prompt = prompt or config.opts.system_prompt
   if prompt ~= "" then
     if type(prompt) == "function" then
       prompt = prompt({
@@ -446,6 +457,14 @@ function Chat:toggle_system_prompt()
   else
     util.notify("Added system prompt")
     self:set_system_prompt()
+  end
+end
+
+---Remove the system prompt from the chat buffer
+---@return nil
+function Chat:remove_system_prompt()
+  if self.messages[1] and self.messages[1].role == config.constants.SYSTEM_ROLE then
+    table.remove(self.messages, 1)
   end
 end
 
@@ -491,7 +510,7 @@ function Chat:add_tool(tool, tool_config)
   end
 
   local id = "<tool>" .. tool .. "</tool>"
-  self.References:add({
+  self.references:add({
     source = "tool",
     name = "tool",
     id = id,
@@ -513,7 +532,7 @@ function Chat:add_tool(tool, tool_config)
 end
 
 ---Add a message to the message table
----@param data table {role: string, content: string}
+---@param data { role: string, content: string }
 ---@param opts? table Options for the message
 ---@return CodeCompanion.Chat
 function Chat:add_message(data, opts)
@@ -578,70 +597,20 @@ function Chat:submit(opts)
 
   local bufnr = self.bufnr
 
-  local message = ts_parse_messages(self, user_role, self.header_line)
+  local message = ts_parse_messages(self, self.header_line)
 
-  for _, ref in ipairs(self.refs) do
-    if ref.bufnr and ref.opts and ref.opts.watched then
-      local changes = self.watchers:get_changes(ref.bufnr)
-      log:debug("Checking watched buffer %d, found %d changes", ref.bufnr, changes and #changes or 0)
+  -- Check if any watched buffers have any changes
+  self.watchers:check_for_changes(self)
 
-      if changes and #changes > 0 then
-        local changes_text = string.format(
-          "Changes detected in `%s` (buffer %d):\n",
-          vim.fn.fnamemodify(api.nvim_buf_get_name(ref.bufnr), ":t"),
-          ref.bufnr
-        )
-
-        for _, change in ipairs(changes) do
-          if change.type == "delete" then
-            changes_text = changes_text
-              .. string.format(
-                "Lines %d-%d were deleted:\n```%s\n%s\n```\n",
-                change.start,
-                change.end_line,
-                vim.bo[ref.bufnr].filetype,
-                table.concat(change.lines, "\n")
-              )
-          elseif change.type == "modify" then
-            changes_text = changes_text
-              .. string.format(
-                "Lines %d-%d were modified from:\n```%s\n%s\n```\nto:\n```%s\n%s\n```\n",
-                change.start,
-                change.end_line,
-                vim.bo[ref.bufnr].filetype,
-                table.concat(change.old_lines, "\n"),
-                vim.bo[ref.bufnr].filetype,
-                table.concat(change.new_lines, "\n")
-              )
-          else -- type == "add"
-            changes_text = changes_text
-              .. string.format(
-                "Lines %d-%d were added:\n```%s\n%s\n```\n",
-                change.start,
-                change.end_line,
-                vim.bo[ref.bufnr].filetype,
-                table.concat(change.lines, "\n")
-              )
-          end
-        end
-
-        self:add_message({
-          role = config.constants.USER_ROLE,
-          content = changes_text,
-        }, { visible = true })
-      end
-    end
-  end
-
-  if not self:has_user_messages(message) or message.content == "" then
+  if not self:has_user_messages(message) then
     return log:warn("No messages to submit")
   end
 
   --- Only send the user's last message if we're not regenerating the response
-  if not opts.regenerate and not vim.tbl_isempty(message) then
+  if not opts.regenerate and not vim.tbl_isempty(message) and message.content ~= "" then
     self:add_message({ role = config.constants.USER_ROLE, content = message.content })
   end
-  message = self.References:clear(self.messages[#self.messages])
+  message = self.references:clear(self.messages[#self.messages])
 
   self:apply_tools_and_variables(message)
   self:check_references()
@@ -667,11 +636,12 @@ function Chat:submit(opts)
   self.current_request = client
     .new({ adapter = settings })
     :request(self.adapter:map_roles(vim.deepcopy(self.messages)), {
-      ---@param err string
+      ---@param err { message: string, stderr: string }
       ---@param data table
       callback = function(err, data)
-        if err then
+        if err and err.stderr ~= "{}" then
           self.status = CONSTANTS.STATUS_ERROR
+          log:error("Error: %s", err.stderr)
           return self:done(output)
         end
 
@@ -736,7 +706,7 @@ function Chat:done(output)
   local assistant_range = self.header_line
   self:set_range(-2)
   self.ui:display_tokens(self.parser, self.header_line)
-  self.References:render()
+  self.references:render()
 
   if self.status == CONSTANTS.STATUS_SUCCESS and self:has_tools() then
     self.tools:parse_buffer(self, assistant_range, self.header_line - 1)
@@ -766,7 +736,7 @@ end
 ---Reconcile the references table to the references in the chat buffer
 ---@return nil
 function Chat:check_references()
-  local refs = self.References:get_from_chat()
+  local refs = self.references:get_from_chat()
   if vim.tbl_isempty(refs) and vim.tbl_isempty(self.refs) then
     return
   end
@@ -857,13 +827,17 @@ function Chat:stop()
     self.current_tool = nil
 
     _G.codecompanion_cancel_tool = true
-    job:shutdown()
+    pcall(function()
+      job:shutdown()
+    end)
   end
   if self.current_request then
     job = self.current_request
     self.current_request = nil
     if job then
-      job:shutdown()
+      pcall(function()
+        job:shutdown()
+      end)
     end
   end
 end
@@ -899,6 +873,8 @@ function Chat:close()
   self = nil
 end
 
+local has_been_reasoning = false
+
 ---Add a message directly to the chat buffer. This will be visible to the user
 ---@param data table
 ---@param opts? table
@@ -907,7 +883,13 @@ function Chat:add_buf_message(data, opts)
   local bufnr = self.bufnr
   local new_response = false
 
-  if (data.role and data.role ~= self.last_role) or (opts and opts.force_role) then
+  local function write(text)
+    for _, t in ipairs(vim.split(text, "\n", { plain = true, trimempty = false })) do
+      table.insert(lines, t)
+    end
+  end
+
+  local function new_role()
     new_response = true
     self.last_role = data.role
     table.insert(lines, "")
@@ -915,13 +897,28 @@ function Chat:add_buf_message(data, opts)
     self.ui:set_header(lines, config.strategies.chat.roles[data.role])
   end
 
-  if data.content then
-    for _, text in ipairs(vim.split(data.content, "\n", { plain = true, trimempty = false })) do
-      table.insert(lines, text)
+  local function append_data()
+    if data.reasoning then
+      has_been_reasoning = true
+      if new_response then
+        table.insert(lines, "### Reasoning")
+        table.insert(lines, "")
+      end
+      write(data.reasoning)
+    else
+      if has_been_reasoning then
+        has_been_reasoning = false
+        table.insert(lines, "")
+        table.insert(lines, "")
+        table.insert(lines, "### Response")
+        table.insert(lines, "")
+      end
+      write(data.content)
     end
+  end
 
+  local function update_buffer()
     self.ui:unlock_buf()
-
     local last_line, last_column, line_count = self.ui:last()
     if opts and opts.insert_at then
       last_line = opts.insert_at
@@ -945,6 +942,17 @@ function Chat:add_buf_message(data, opts)
       self.ui:follow()
     end
   end
+
+  -- Handle a new role
+  if (data.role and data.role ~= self.last_role) or (opts and opts.force_role) then
+    new_role()
+  end
+
+  -- Append the output from the LLM
+  if data.content or data.reasoning then
+    append_data()
+    update_buffer()
+  end
 end
 
 ---When a request has finished, reset the chat buffer
@@ -961,7 +969,7 @@ function Chat:get_messages()
 end
 
 ---Subscribe to a chat buffer
----@param event table {name: string, type: string, callback: fun}
+---@param event { name: string, type: string, callback: fun() }
 function Chat:subscribe(event)
   table.insert(self.subscribers, event)
 end